22113 change thumbnail generation
Move handleThumbNailGeneration to ObjectStoreMetaDataService.java
