+++ src/main/java/ca/gc/aafc/objectstore/api/file/FileController.java
@@ -6,7 +6,7 @@ import ca.gc.aafc.objectstore.api.entities.ObjectStoreMetadata;
 import ca.gc.aafc.objectstore.api.entities.ObjectUpload;
 import ca.gc.aafc.objectstore.api.exif.ExifParser;
 import ca.gc.aafc.objectstore.api.minio.MinioFileService;
-import ca.gc.aafc.objectstore.api.service.DerivativeService;
+import ca.gc.aafc.objectstore.api.service.ObjectStoreMetaDataService;
 import ca.gc.aafc.objectstore.api.service.ObjectStoreMetadataReadService;
 import ca.gc.aafc.objectstore.api.service.ObjectUploadService;
 import io.crnk.core.exception.UnauthorizedException;
@@ -61,7 +61,7 @@ public class FileController {
   private static final int READ_AHEAD_BUFFER_SIZE = 10 * 1024;
 
   private final ObjectUploadService objectUploadService;
-  private final DerivativeService derivativeService;
+  private final ObjectStoreMetaDataService objectStoreMetaDataService;
   private final MinioFileService minioService;
   private final ObjectStoreMetadataReadService objectStoreMetadataReadService;
   private final MediaTypeDetectionStrategy mediaTypeDetectionStrategy;
@@ -74,7 +74,7 @@ public class FileController {
   public FileController(
     MinioFileService minioService,
     ObjectUploadService objectUploadService,
-    DerivativeService derivativeService,
+    ObjectStoreMetaDataService objectStoreMetaDataService,
     ObjectStoreMetadataReadService objectStoreMetadataReadService,
     MediaTypeDetectionStrategy mediaTypeDetectionStrategy,
     DinaAuthenticatedUser authenticatedUser,
@@ -86,7 +86,7 @@ public class FileController {
     this.mediaTypeDetectionStrategy = mediaTypeDetectionStrategy;
     this.authenticatedUser = authenticatedUser;
     this.messageSource = messageSource;
-    this.derivativeService = derivativeService;
+    this.objectStoreMetaDataService = objectStoreMetaDataService;
   }
 
   @PostMapping("/file/{bucket}/derivative")
@@ -171,8 +171,10 @@ public class FileController {
     @PathVariable String bucket,
     @PathVariable UUID fileId
   ) throws IOException {
-    Derivative derivative = derivativeService.findByFileId(fileId)
-      .orElseThrow(() -> buildNotFoundException(bucket, fileId));
+    Derivative derivative = objectStoreMetaDataService
+      .findAll(Derivative.class, (criteriaBuilder, root) -> new Predicate[]{
+        criteriaBuilder.equal(root.get("fileIdentifier"), fileId)}, null, 0, 1)
+      .stream().findFirst().orElseThrow(() -> buildNotFoundException(bucket, fileId));
     String fileName = derivative.getFileIdentifier() + derivative.getFileExtension();
     return download(bucket, fileId, fileName, true);
   }
@@ -187,10 +189,10 @@ public class FileController {
       .loadObjectStoreMetadataByFileId(fileId)
       .orElseThrow(() -> buildNotFoundException(bucket, fileId));
 
-    Derivative derivative = derivativeService.findAll(Derivative.class,
+    Derivative derivative = objectStoreMetaDataService.findAll(Derivative.class,
       (criteriaBuilder, root) -> new Predicate[]{
         criteriaBuilder.equal(root.get("acDerivedFrom"), objectStoreMetadata),
-        criteriaBuilder.equal(root.get("objectSubtype"), derivativeService.getThumbNailSubType())},
+        criteriaBuilder.equal(root.get("objectSubtype"), objectStoreMetaDataService.getThumbNailSubType())},
       null, 0, 1).stream().findFirst()
       .orElseThrow(() -> buildNotFoundException(bucket, fileId));
     return downloadDerivative(bucket, derivative.getFileIdentifier());
